Renamed Value to Item
This is because I realised we're about to write a load of code that will have Value#value and that gives me the ick.

Also there's now a symmetry between definitions and items.

There are DataStructures::Containers and DataStructures::Definition::Containers.  There are DataStructures::Items and DataStructures::Definition::Items.  In each case the latter is the part of the template the former holds the actual data.