Creating items for storing data - including RepeatingGroups
Definitions: RepeatingGroups contain a single Repeat that in turn contains the data structure definitions.

Items: RepeatingGroups create a single Repeat that in turn holds the data items.

However, on the item side of things, we are about to add in the ability to add in new repeats - so each repeating group can have multiple repeating blocks of data.