Validation rules
The real world strikes again.

Rails forms often send empty strings for field values.  And adding allow_nil on the validation deals with this, because Rails knows its going to an integer field, so becomes nil.

But because the value of a field is dynamic, Rails can't do that, so an empty string is triggering the validations, even if the field is not required.

So instead we have to check for blank non-required fields and skip the validation in those cases