Moved container
It was bugging me that we had app/models/data_structures/field and then app/models/concerns/data_structures/container AND lib/data_structures/whatever

Field is a namespaced active record model so belongs in app/models/data_structures.
But container is just a module, so it can safely live in lib and we end up with less folders to search for stuff in.