Templates do not return a path_name
This means that items that are directly to the template will have a root path

Previously a top-level text_field would have a path of "1/what_is_your_name" whereas now it will be "what_is_your_name" - which makes a bit more sense.

Unfortunately there's a nasty if statement in the spec for it though.  I'll sort that another time.